+++ katalog-server/src/main/kotlin/com/bol/katalog/api/v1/SchemaResource.kt
@@ -1,12 +1,18 @@
 package com.bol.katalog.api.v1
 
-import com.bol.katalog.api.*
+import com.bol.katalog.api.PaginationRequest
+import com.bol.katalog.api.SortingRequest
+import com.bol.katalog.api.paginate
+import com.bol.katalog.api.sort
+import com.bol.katalog.cqrs.Aggregate
 import com.bol.katalog.features.registry.*
+import com.bol.katalog.security.PermissionManager
 import com.bol.katalog.security.monoWithUserDetails
 import com.bol.katalog.users.GroupPermission
 import org.springframework.http.HttpStatus
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.web.bind.annotation.*
+import org.springframework.web.server.ResponseStatusException
 import java.time.Instant
 import java.util.*
 
@@ -14,8 +20,8 @@ import java.util.*
 @RequestMapping("/api/v1/schemas")
 @PreAuthorize("hasRole('USER')")
 class SchemaResource(
-    private val registry: RegistryAggregate,
-    private val permissionChecker: PermissionChecker
+    private val registry: Aggregate<RegistryState>,
+    private val permissionManager: PermissionManager
 ) {
     object Responses {
         data class Schema(
@@ -73,13 +79,10 @@ class SchemaResource(
 
     private suspend fun toResponse(it: Schema): Responses.Schema {
         return registry.read {
-            val namespaceId = getSchemaNamespaceId(it.id)
-            val namespace = getNamespace(namespaceId)
-
             Responses.Schema(
                 id = it.id,
                 createdOn = it.createdOn,
-                namespace = Responses.Schema.Namespace(namespace.id, namespace.name),
+                namespace = Responses.Schema.Namespace(it.namespace.id, it.namespace.name),
                 schema = it.name
             )
         }
@@ -103,7 +106,6 @@ class SchemaResource(
     fun create(
         @RequestBody data: Requests.NewSchema
     ) = monoWithUserDetails {
-        permissionChecker.requireNamespace(data.namespaceId, GroupPermission.CREATE)
         val id: SchemaId = UUID.randomUUID().toString()
         registry.send(CreateSchemaCommand(data.namespaceId, id, data.schema, SchemaType.default()))
         Responses.SchemaCreated(id)
@@ -114,7 +116,11 @@ class SchemaResource(
     fun delete(
         @PathVariable id: SchemaId
     ) = monoWithUserDetails {
-        permissionChecker.requireSchema(id, GroupPermission.DELETE)
+        val schema = registry.read { getSchema(id) }
+        permissionManager.requirePermission(schema.namespace.groupId, GroupPermission.DELETE) {
+            throw ResponseStatusException(HttpStatus.FORBIDDEN)
+        }
+
         registry.send(DeleteSchemaCommand(id))
     }
 }
